requirements.txt
APScheduler==3.6.3
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
config==0.5.0.post0
cryptography==3.4.7
idna==2.10
paramiko==2.7.2
pycparser==2.20
PyNaCl==1.4.0
python-decouple==3.4
python-telegram-bot==13.4.1
aiogram==2.13
pytz==2021.1
requests==2.25.1
six==1.15.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.4
